+++ src/rngs/mock.rs
@@ -10,6 +10,9 @@
 
 use rand_core::{impls, Error, RngCore};
 
+#[cfg(feature = "serde1")]
+use serde::{Serialize, Deserialize};
+
 /// A simple implementation of `RngCore` for testing purposes.
 ///
 /// This generates an arithmetic sequence (i.e. adds a constant each step)
@@ -25,6 +28,7 @@ use rand_core::{impls, Error, RngCore};
 /// assert_eq!(sample, [2, 3, 4]);
 /// ```
 #[derive(Debug, Clone)]
+#[cfg_attr(feature = "serde1", derive(Serialize, Deserialize))]
 pub struct StepRng {
     v: u64,
     a: u64,
@@ -63,5 +67,20 @@ impl RngCore for StepRng {
     fn try_fill_bytes(&mut self, dest: &mut [u8]) -> Result<(), Error> {
         self.fill_bytes(dest);
         Ok(())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    #[cfg(feature = "serde1")]
+    fn test_serialization_step_rng() {
+        let some_rng = StepRng::new(42, 7);
+        let de_some_rng: StepRng = bincode::deserialize(&bincode::serialize(&some_rng).unwrap()).unwrap();
+        assert_eq!(some_rng.v, de_some_rng.v);
+        assert_eq!(some_rng.a, de_some_rng.a);
+
     }
 }